feat(Facade): add ability to restrict pointer by target point
The Zinnia PointsCast now allows to restrict the target point of a
cast to allow or deny certain points of space from being valid.

This is now used with the PointerFacade to allow a pointer to determine
if the point in which the cursor is colliding with is valid based on
a Rule.

It can be used in conjunction with the NavMeshRule to determine if the
pointer is within the bounds of a NavMesh to consider it a valid target
to select.