fix(Configuration): ensure setup is run correctly on enable
The RunWhenActiveAndEnabled logic could cause an issue where it was
hitting race conditions and not setting up correctly. This is now
handled by calling a single method rather than queuing methods.

The configuration now also disables the points caster instead of the
moment processor that controls the points caster, so the relevant
setup is run correctly whenever the pointer is activated.

The Facade and Configuration reference fields are also now public
to allow manual creation if required.